修改Date转json之后由时区产生的错误，改成使用LocalDate

+++ src/main/java/com/donglu/controller/VisitorController.java
@@ -5,14 +5,15 @@ import com.donglu.config.SecurityInterceptor;
 import com.donglu.mapper.LoginUserMapper;
 import com.donglu.mapper.VisitorControlMapper;
 import com.github.pagehelper.PageHelper;
-import com.google.common.collect.Maps;
+import com.google.common.base.Strings;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.propertyeditors.CustomDateEditor;
-import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
 
-import java.text.SimpleDateFormat;
-import java.util.*;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 /**
  * 访客控制器
@@ -32,6 +33,9 @@ public class VisitorController {
     @RequestMapping(value = "/login",method = RequestMethod.POST)
     @NoSecurity
     public Response login(@RequestParam String accountName, @RequestParam String accountPassword){
+        if (Strings.isNullOrEmpty(accountName) || Strings.isNullOrEmpty(accountPassword)) {
+            return new Response().failureMsg("用户名与密码不可为空");
+        }
         CardUser cardUser = loginUserMapper.findCardUser(accountName);
         Optional<Response> response = checkLogin(cardUser, accountPassword);
         if (response.isPresent()) {
@@ -54,6 +58,9 @@ public class VisitorController {
     @RequestMapping(value = "/login",method = RequestMethod.PUT)
     @NoSecurity
     public Response modifyPassword(@RequestParam String loginName, @RequestParam String loginPassword,@RequestParam String newPassword){
+        if (Strings.isNullOrEmpty(loginName) || Strings.isNullOrEmpty(loginPassword) || Strings.isNullOrEmpty(newPassword)) {
+            return new Response().failureMsg("用户名与密码不可为空");
+        }
         CardUser cardUser = loginUserMapper.findCardUser(loginName);
         Optional<Response> response = checkLogin(cardUser, loginPassword);
         if (response.isPresent()) {
@@ -90,6 +97,16 @@ public class VisitorController {
         CardUser cardUser = (CardUser) securityInterceptor.getLogin();
         conditionMap.put("userName", cardUser.getUserName());
 
+        Optional.ofNullable(conditionMap.get("visitorTime")).ifPresent(s->{
+            String visitorTime = (String) s;
+            String[] split = visitorTime.split(" - ");
+            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
+            String startTime = LocalDate.parse(split[0].trim()).atTime(0, 0, 0).format(formatter);
+            String endTime = LocalDate.parse(split[1].trim()).atTime(23, 58, 59).format(formatter);
+            conditionMap.put("visitorStartTime",startTime);
+            conditionMap.put("visitorEndTime",endTime);
+        });
+
         List<VisitorBooking> visitorBookList = visitorControlMapper.findVisitorBookList(conditionMap);
         Long count = visitorControlMapper.countVisitorBookList(conditionMap);
         return new ResponseTableList(count, visitorBookList);
@@ -97,31 +114,22 @@ public class VisitorController {
 
     @RequestMapping(value = "/visitor",method = RequestMethod.POST)
     public Response insertVisitorBooking(VisitorBooking visitorBooking) {
-        HashMap<String, Object> paramMap = Maps.newHashMap();
-        paramMap.put("IDCard",visitorBooking.getIDCard());
-        Long aLong = visitorControlMapper.countVisitorBookList(paramMap);
-        if (aLong > 0) {
-            return new Response().failureMsg("该访客身份证件己登记");
-        }
         return insertOrUpdateVisitorBooking(visitorBooking);
     }
 
     @RequestMapping(value = "/visitor",method = RequestMethod.PUT)
     public Response updateVisitorBooking(VisitorBooking visitorBooking) {
-        HashMap<String, Object> paramMap = Maps.newHashMap();
-        paramMap.put("IDCard",visitorBooking.getIDCard());
-        paramMap.put("id",visitorBooking.getId());
-        Long aLong = visitorControlMapper.countVisitorBookList(paramMap);
-        if (aLong > 0) {
-            return new Response().failureMsg("该访客身份证件己登记");
-        }
-
         return insertOrUpdateVisitorBooking(visitorBooking);
     }
 
     @RequestMapping(value = "/visitor/{id}",method = RequestMethod.DELETE)
     public Response deleteVisitorBooking(@PathVariable Long id) {
         try {
+            VisitorBooking visitor = visitorControlMapper.findVisitorBooking(id);
+            if (visitor != null && "已完成".equals(visitor.getStatus())) {
+                return new Response().failureMsg("己完成的预约不能再进行删除");
+            }
+
             Long count = visitorControlMapper.deleteVisitorBooking(id);
             if (count > 0) {
                 return new Response().successMsg("删除成功");
@@ -141,19 +149,23 @@ public class VisitorController {
 
         CardUser cardUser = (CardUser) obj;
         visitorBooking.copyFromCardUser(cardUser);
+
+//        visitorBooking.setVisitorTime(visitorBooking.getVisitorTime().);
         if (visitorBooking.getId() == null) {
+            visitorBooking.setStatus("未完成");
             visitorControlMapper.insert(visitorBooking);
             return new Response().successMsg("添加成功");
         } else {
+            VisitorBooking visitor = visitorControlMapper.findVisitorBooking(visitorBooking.getId());
+            if (visitor != null && "已完成".equals(visitor.getStatus())) {
+                return new Response().failureMsg("己完成的预约不能再进行修改");
+            }
+
+            visitorBooking.setStatus("未完成");
             visitorControlMapper.update(visitorBooking);
             return new Response().successMsg("修改成功");
         }
 
     }
 
-    @InitBinder
-    protected void initBinder(WebDataBinder binder) {
-        SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
-        binder.registerCustomEditor(Date.class, new CustomDateEditor(dateFormat, true));
-    }
 }
